extract proton and neutron
